Title: 컴퓨터와 함께하는 숫자 게임 2, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240821/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -44,7 +44,7 @@ public class Main {
         a = sc.nextLong();
         b = sc.nextLong();
 
-        for (int i = a; i <= b; i++) {
+        for (long i = a; i <= b; i++) {
             find(i);
         }
         System.out.print(minCount + " " + maxCount);
